+++ lib/presenters/model.rb
@@ -1,0 +1,8 @@
+module Presenters
+  module Model
+    def presenter
+      @presenter = Helper.present(self) if @presenter.nil?
+      @presenter
+    end
+  end
+end