+++ lib/presenters/presenter.rb
@@ -1,4 +1,5 @@
 module Presenters
   class Presenter < SimpleDelegator
+    include Helper
   end
 end