+++ spec/model_spec.rb
@@ -1,0 +1,10 @@
+RSpec.describe Presenters::Model do
+  let(:post) { Post.new }
+  subject { post.presenter }
+
+  describe "#presenter" do
+    context "without block" do
+      its(:class) { should eq PostPresenter }
+    end
+  end
+end