+++ spec/support/post_presenter.rb
@@ -2,4 +2,12 @@ class PostPresenter < Presenter
   def title
     super || 'No Title'
   end
+
+  def poster
+    if user.nil?
+      'No Poster'
+    else
+      present(user).name
+    end
+  end
 end