+++ spec/support/user_presenter.rb
@@ -1,0 +1,5 @@
+class UserPresenter < Presenter
+  def name
+    super || 'No Name'
+  end
+end