+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.2.0 / 2015-08-23
+
+* Include helpers in Presenter.
+* Add presenter method to ActiveRecord:Base.
+* Add presenters method.
+
 # v0.1.0 / 2015-08-23
 
 Create project